Canteen API finished and pass the test
JSON格式
canteen/info
{
    "username": "master",
    "canteenName": "一餐厅"
}

canteen/create:
{
    "username": "master",
    "canteenName": "四餐厅",
    "location": "杨浦区军工路516号上海理工大学内"
}
{
    "username": "master",
    "canteenName": "666餐厅",
    "location": "杨浦区军工路516号上海理工大学内"
}

canteen/canteen-name:
{
    "username": "master",
    "id": 7,
    "canteenName": "上海理工大学肠胃科"
}

canteen/intro:
{
    "username": "master",
    "canteenName": "四餐厅",
    "intro": "我才是基础学院学生的唯二选择"
}

canteen/location:
{
    "username": "master",
    "canteenName": "上海理工大学肠胃科",
    "location": "上海理工大学"
}

canteen/business-hours:
{
    "username": "master",
    "canteenName": "四餐厅",
    "businessHours": "06:00 - 22:00"
}

canteen/announcement:
{
    "username": "master",
    "canteenName": "四餐厅",
    "announcement": "基础学院的荣耀就由我来守护"
}

canteen/delete:
{
    "username": "master",
    "canteenName": "上海理工大学肠胃科"
}

+++ spring-boot/src/main/java/com/windsnow1025/canteenmanagement/springboot/logic/CanteenLogic.java
@@ -4,6 +4,8 @@ import com.windsnow1025.canteenmanagement.springboot.dao.CanteenDAO;
 import com.windsnow1025.canteenmanagement.springboot.model.Canteen;
 import com.windsnow1025.canteenmanagement.springboot.util.JwtUtil;
 
+import java.util.Objects;
+
 public class CanteenLogic {
     private final CanteenDAO canteenDAO;
 
@@ -11,42 +13,69 @@ public class CanteenLogic {
         canteenDAO = new CanteenDAO();
     }
 
-    public Canteen getInfo(String token){
-        String canteenName = JwtUtil.parseJWT(token);
-        return canteenDAO.selectCanteenByCanteenName(canteenName);
+    public Canteen getInfo(String token, String username, String canteenName){
+        if (Objects.equals(username, JwtUtil.parseJWT(token))) {
+            return canteenDAO.selectCanteenByCanteenName(canteenName);
+        }else
+        {
+            return null;
+        }
     }
 
-    public boolean insert(Canteen canteen){
-        return canteenDAO.insert(canteen);
+    public boolean insert(String token, String username, String canteenName, String intro, String location, String businessHours, String announcement){
+        if (Objects.equals(username, JwtUtil.parseJWT(token))) {
+            Canteen canteen = new Canteen(canteenName, intro, location, businessHours, announcement);
+            return canteenDAO.insert(canteen);
+        }else {
+            return false;
+        }
     }
 
-    public boolean updateCanteenName(String token, String newCanteenName){
-        int id = Integer.parseInt(JwtUtil.parseJWT(token));
-        return canteenDAO.updateCanteenName(id, newCanteenName);
+    public boolean updateCanteenName(String token, String username, int id, String newCanteenName){
+        if (Objects.equals(username, JwtUtil.parseJWT(token))) {
+            return canteenDAO.updateCanteenName(id, newCanteenName);
+        }else {
+            return false;
+        }
     }
 
-    public boolean updateIntro(String token, String newIntro){
-        String canteenName = JwtUtil.parseJWT(token);
-        return canteenDAO.updateIntro(canteenName, newIntro);
+    public boolean updateIntro(String token, String username, String canteenName, String newIntro){
+        if (Objects.equals(username, JwtUtil.parseJWT(token))){
+            return canteenDAO.updateIntro(canteenName, newIntro);
+        }else {
+            return false;
+        }
     }
 
-    public boolean updateLocation(String token, String newLocation){
-        String canteenName = JwtUtil.parseJWT(token);
-        return canteenDAO.updateLocation(canteenName, newLocation);
+    public boolean updateLocation(String token, String username, String canteenName, String newLocation){
+        if (Objects.equals(username, JwtUtil.parseJWT(token))){
+            return canteenDAO.updateLocation(canteenName, newLocation);
+        }else {
+            return false;
+        }
     }
 
-    public boolean updateBusinessHour(String token, String newBusinessHours){
-        String canteenName = JwtUtil.parseJWT(token);
-        return canteenDAO.updateBusinessHour(canteenName, newBusinessHours);
+    public boolean updateBusinessHour(String token, String username, String canteenName, String newBusinessHours){
+        if (Objects.equals(username, JwtUtil.parseJWT(token))){
+            return canteenDAO.updateBusinessHour(canteenName, newBusinessHours);
+        }else {
+            return false;
+        }
     }
 
-    public boolean updateAnnouncement(String token, String newAnnouncement){
-        String canteenName = JwtUtil.parseJWT(token);
-        return canteenDAO.updateAnnouncement(canteenName, newAnnouncement);
+    public boolean updateAnnouncement(String token, String username, String canteenName, String newAnnouncement){
+        if (Objects.equals(username, JwtUtil.parseJWT(token))){
+            return canteenDAO.updateAnnouncement(canteenName, newAnnouncement);
+        }else {
+            return false;
+        }
     }
 
-    public boolean delete(String token){
-        String canteenName = JwtUtil.parseJWT(token);
-        return canteenDAO.delete(canteenName);
+    public boolean delete(String token, String username, String canteenName){
+        if (Objects.equals(username, JwtUtil.parseJWT(token))){
+            return canteenDAO.delete(canteenName);
+        }else {
+            return false;
+        }
     }
 }
